Optimize dragging the canvas with the mouse to do less work per frame
When moving the canvas with the mouse, each frame of animation was
causing all nodes and links to rerender.  On large diagrams, this
can feel quite slow and cause low fps while navigating around the
diagram.

By setting an additional flag indicating when we're dragging, we can
stop the contents of the LinkLayerWidget and the NodeLayerWidget
from re-rendering.

This uses <React.Fragment>, which requires React 16.2 or later.